Fix delete key when focused on a form

+++ src/routes/route_check/jat_check/EditJunction.svelte
@@ -64,6 +64,12 @@
       e.stopPropagation();
       editing = null;
     } else if (editing != null && e.key == "Delete") {
+      // Let the delete key work in forms
+      let tag = (e.target as HTMLElement).tagName;
+      if (tag == "INPUT" || tag == "TEXTAREA") {
+        return;
+      }
+      e.stopPropagation();
       deleteItem();
     }
   }
